[💎 : refactor] 양 사이드 창 컴포넌트 분리
- 추후 프로필 및 추천 사용자 영역 작업을 위해 컴포넌트를 나누어 관리하는 것이 편할 것으로 판단
