+++ src/components/right-side-menu/search-keyword.tsx
@@ -1,13 +1,11 @@
 import React, { useState } from 'react';
 import { useSetRecoilState } from 'recoil';
 import { useNavigate } from 'react-router-dom';
-import WindowTop from '@compo/window-top.tsx';
 import searchKeywordAtom from '@atom/search-keyword.tsx';
-import * as W from '@style/window.ts';
-import * as S from '@style/search.ts';
+import * as S from '@style/search-input.ts';
 import { ReactComponent as IconSearch } from '@img/i-search.svg';
 
-export default function Search() {
+export default function SearchKeyword() {
   const navigate = useNavigate();
   const [searchValue, setSearchValue] = useState('');
   const setSearchKeyword = useSetRecoilState(searchKeywordAtom);
@@ -21,22 +19,19 @@ export default function Search() {
     navigate(`/search?query=${searchValue}`);
   };
   return (
-    <W.Window>
-      <WindowTop />
-      <S.Form onSubmit={onSubmit}>
-        <S.FormInput
-          onChange={onChange}
-          name="searchKeyword"
-          value={searchValue}
-          placeholder="검색어 입력"
-          type="text"
-          required
-        ></S.FormInput>
-        <S.FormButton type="submit">
-          <span className="a11yHidden">검색하기</span>
-          <IconSearch />
-        </S.FormButton>
-      </S.Form>
-    </W.Window>
+    <S.Form onSubmit={onSubmit}>
+      <S.FormInput
+        onChange={onChange}
+        name="searchKeyword"
+        value={searchValue}
+        placeholder="검색어 입력"
+        type="text"
+        required
+      ></S.FormInput>
+      <S.FormButton type="submit">
+        <span className="a11yHidden">검색하기</span>
+        <IconSearch />
+      </S.FormButton>
+    </S.Form>
   );
 }
