+++ src/styles/mini-profile.ts
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+import { blackColor } from '@style/global.ts';
+
+export const MiniProfile = styled.div`
+  width: 100%;
+  height: 200px;
+  margin-top: auto;
+  background-color: ${blackColor};
+`;
+
+export const dummy = styled.div`
+  display: none;
+`;
