+++ src/styles/navigation.ts
