+++ src/styles/search-input.ts
